set version numbers in requirements.txt and fix ModelSchema in location.py
This isn't full version pinning, but should at least keep our dependencies compatible, it will need pinning later to ensure that transitive dependencies are also compatible.
There was a deprecation in one of the SQLAlchemy libraries that is in use, this is also fixed in this commit by updating to the new suggested class.
See https://flask-marshmallow.readthedocs.io/en/latest/changelog.html#id2